fix: render share template search params client side
